[DI-1370]  After running the transformLoad delete the files on the DataImport folder with action (Always Delete) (#113)
Updata workflow to delete files from the dataimport folder.
Files are removed from the data import folder, and they are not removed from the FTP Server (external source)

+++ DataImport.Common/LocalFileService.cs
@@ -97,8 +97,12 @@ namespace DataImport.Common
 
         public Task Delete(File file)
         {
-            System.IO.File.Delete(new Uri(file.Url).LocalPath);
-            return CleanAgentDirectoryIfEmpty(file);
+            if (System.IO.File.Exists(new Uri(file.Url).LocalPath))
+            {
+                System.IO.File.Delete(new Uri(file.Url).LocalPath);
+                return CleanAgentDirectoryIfEmpty(file);
+            }
+            return Task.CompletedTask;
         }
 
         private static Task CleanAgentDirectoryIfEmpty(File file)
